ir3: Simplify and rationalize shading rate LUT
We had an extra 16 entries in the VK-to-HW table that were clearly
unnecessary because Vulkan does not allow values greater than 16 for the
primitive shading rate. This appears to be an extra debug/test thing
added by the blob. Similarly there were unused entries in the HW-to-VK
table that shouldn't be necessary. Delete them.

The HW-to-VK table was also inconsistent about whether invalid values
should be 0 or 11, fix that too.

Part-of: <https://gitlab.freedesktop.org/mesa/mesa/-/merge_requests/36892>