pan/lib: clean up tilebuffer size helpers
This way we report the *actual* tilebuffer sizes rather than reporting
twice the amount and then correcting for it later on. The distinction
really only matters on V4 and V6 GPUs, but let's get it correct.

This shouln't have any functional changes, it just makes it easier to
reason about what each GPU supports.

While we're at it, correct the typo in the comment about what gens
supports disabling pipelining. It's *not* supported on V6.

Reviewed-by: Eric R. Smith <eric.smith@collabora.com>
Reviewed-by: Boris Brezillon <boris.brezillon@collabora.com>
Part-of: <https://gitlab.freedesktop.org/mesa/mesa/-/merge_requests/35755>