util/u_queue: Fix data race on num_threads during finish.
This should fix a deadlock I saw on
dEQP-EGL.functional.sharing.gles2.multithread.random.programs.link.7 on
radeonsi, where we were waiting on an invalid fence[1] value.  This was
probably because between when we started setting up the fences for
u_queue_finish and when we waited on those fences, a second thread was
created.

Closes: #13738
Part-of: <https://gitlab.freedesktop.org/mesa/mesa/-/merge_requests/36919>