nil: Delete some useless image alignment code
This was copied+pasted from the old GL driver but we never use any of
those PTE kinds since they're only for compressed depth pre-Turing.
We've also never provied that this is actually good for anything.  Just
delete it for now.

Part-of: <https://gitlab.freedesktop.org/mesa/mesa/-/merge_requests/37094>