ir3: use shared masks for cov when scalar ALU is supported
Whenever cov doesn't work and we have to use a mask instead, the mask
was created as a non-shared immediate, preventing the use of the scalar
ALU for the and instruction. Fix this by creating a shared immediate
when the scalar ALU is supported.

Signed-off-by: Job Noorman <jnoorman@igalia.com>
Part-of: <https://gitlab.freedesktop.org/mesa/mesa/-/merge_requests/37104>