st/glsl: encapsulate more in st_nir_state_variable_create()
This will allow us to fix bugs with driver_location in following
patches while keeping code to a minimum.

For now we always set uniform packing to false, this will be
corrected as we go in following patches for now it doesn't matter
as the driver location is later overwritten.

Reviewed-by: Mike Blumenkrantz <michael.blumenkrantz@gmail.com>
Part-of: <https://gitlab.freedesktop.org/mesa/mesa/-/merge_requests/37037>