Revert "ci: Update CrosVM and Virglrenderer"
This reverts commit ab5605aab3241c87b3828abcda861efa58d20ae8.

We require 1.78 on 25.2 and we should CI against 1.78 on 25.2.

If something in CI requires a newer version, this should be handled
differently. If we require a newer rustc version because of a dependency
we should properly bump the rustc version requirement.

But please no random bumps like this because we don't want to cause random
build failures due to changes in rustc.

Part-of: <https://gitlab.freedesktop.org/mesa/mesa/-/merge_requests/37198>