zink/ci: drop gbm override now that debian has a usable xorg
https://gitlab.freedesktop.org/mesa/mesa/-/merge_requests/35853 gave us
a debian package which has `--disable-thread-safety-constructor` removed,
fixing this bug.

The `blender-demo-ellie_pose.trace` and `counterstrike-v2.trace` updates
are because there's a rectangle at the top (for blender, window title
bar I expect) or on the sides (counterstrike, padding for image ratio)
which renders grey in gbm and black in xorg.

Part-of: <https://gitlab.freedesktop.org/mesa/mesa/-/merge_requests/37437>